test: move the batch model_group regression into test_router.py
CLAUDE.md asks bug fixes to extend the existing mapped test file rather than
add a new one, and tests/test_litellm/test_router.py already covers
Router.aretrieve_batch. Fold the two cases in next to that coverage and drop
the standalone file.

The helpers are prefixed so they read unambiguously in a shared file, and the
respx context stays open while the payload is awaited, since the usage
accounting reads the batch's output file from the success handler.
